+++ Core/include/Acts/Detector/detail/SupportSurfacesHelper.hpp
@@ -32,7 +32,8 @@ using namespace UnitLiterals;
 namespace Experimental::detail::SupportSurfacesHelper {
 
 using SupportSurfaceComponents =
-    std::tuple<Surface::SurfaceType, std::vector<double>, Acts::Transform3>;
+    std::tuple<Surface::SurfaceType, std::vector<long double>,
+               Acts::Transform3>;
 
 /// @brief  A support creator turns an extend into a vector of bound values
 using SurfaceComponentsCreator =
@@ -44,15 +45,15 @@ struct CylindricalSupport {
   /// - negative indicates inner support
   /// - zero is centered (not recommended)
   /// - positive indicates outer support
-  double rOffset = 0.;
+  long double rOffset = 0.;
 
   /// Clearance in z in order to make the support surfaces
   /// not touch the volume boundaries
-  std::array<double, 2u> zClearance = {1_mm, 1_mm};
+  std::array<long double, 2u> zClearance = {1_mm, 1_mm};
 
   /// Clearance in phi if a sectoral support is chosen
   /// not to touch the volume boundaries
-  std::array<double, 2u> phiClearance = {0.0001_rad, 0.0001_rad};
+  std::array<long double, 2u> phiClearance = {0.0001_rad, 0.0001_rad};
 
   // Type is obviously a cylinder
   static constexpr Surface::SurfaceType type = Surface::SurfaceType::Cylinder;
@@ -71,15 +72,15 @@ struct DiscSupport {
   /// - negative indicates support an z min
   /// - zero is centered
   /// - positive indicates support at z max
-  double zOffset = 0.;
+  long double zOffset = 0.;
 
   /// Clearance in r in order to make the support surfaces
   /// not touch the volume boundaries
-  std::array<double, 2u> rClearance = {1_mm, 1_mm};
+  std::array<long double, 2u> rClearance = {1_mm, 1_mm};
 
   /// Clearance in phi if a sectoral support is chosen
   /// not to touch the volume boundaries
-  std::array<double, 2u> phiClearance = {0.0001_rad, 0.0001_rad};
+  std::array<long double, 2u> phiClearance = {0.0001_rad, 0.0001_rad};
 
   // Type is obviously a disc
   static constexpr Surface::SurfaceType type = Surface::SurfaceType::Disc;
@@ -98,14 +99,14 @@ struct RectangularSupport {
   AxisDirection pPlacement = AxisDirection::AxisZ;
 
   /// Offset in position placement
-  double pOffset = 0.;
+  long double pOffset = 0.;
 
   /// Clearance in first local direction - cyclic order
-  std::array<double, 2u> loc0Clearance = {1_mm, 1_mm};
+  std::array<long double, 2u> loc0Clearance = {1_mm, 1_mm};
 
   /// Clearance in phi if a sectoral support is chosen
   /// not to touch the volume boundaries
-  std::array<double, 2u> loc1Clearance = {1_mm, 1_mm};
+  std::array<long double, 2u> loc1Clearance = {1_mm, 1_mm};
 
   // Type is obviously a plane
   static constexpr Surface::SurfaceType type = Surface::SurfaceType::Plane;
